+++ ckanext/dcatde/harvesters/dcatde_rdf.py
@@ -1,6 +1,9 @@
 #!/usr/bin/python
 # -*- coding: utf8 -*-
 import time
+import json
+import logging
+import pylons
 
 from ckan import plugins as p
 from ckan import model
@@ -8,21 +11,20 @@ from ckanext.dcat.harvesters.rdf import DCATRDFHarvester
 from ckanext.dcat.interfaces import IDCATRDFHarvester
 from ckanext.dcatde.dataset_utils import set_extras_field
 from ckanext.dcatde.harvesters.harvest_utils import HarvestUtils
-from ckanext.harvest.harvesters import HarvesterBase
 from ckanext.harvest.model import HarvestObject, HarvestObjectExtra
-import json
-import logging
 
 LOGGER = logging.getLogger(__name__)
 
 CONFIG_PARAM_HARVESTED_PORTAL = 'harvested_portal'
+CONFIG_PARAM_RESOURCES_REQUIRED = 'resources_required'
 EXTRA_KEY_HARVESTED_PORTAL = 'metadata_harvested_portal'
+RES_EXTRA_KEY_LICENSE = 'license'
 
 
 class DCATdeRDFHarvester(DCATRDFHarvester):
 
     p.implements(IDCATRDFHarvester)
-    
+
     # -- begin IDCATRDFHarvester implementation --
     def before_download(self, url, harvest_job):
         return url, []
@@ -60,6 +62,17 @@ class DCATdeRDFHarvester(DCATRDFHarvester):
             return json.loads(source_config).get(CONFIG_PARAM_HARVESTED_PORTAL)
 
         return ''
+
+    def _get_resources_required_config(self, source_config):
+        if source_config:
+            return json.loads(source_config).get(CONFIG_PARAM_RESOURCES_REQUIRED, False)
+
+        return False
+
+    def _get_fallback_license(self):
+        fallback = pylons.config.get('ckanext.dcatde.harvest.default_license',
+                                     'http://dcat-ap.de/def/licenses/other-closed')
+        return fallback
 
     def _mark_datasets_for_deletion(self, guids_in_source, harvest_job):
         # This is the same as the method in the base class, except that a different query is used.
@@ -104,7 +117,12 @@ class DCATdeRDFHarvester(DCATRDFHarvester):
         guids_in_db = guid_to_package_id.keys()
 
         # Get objects/datasets to delete (ie in the DB but not in the source)
-        guids_to_delete = set(guids_in_db) - set(guids_in_source)
+        guids_in_source_unique = set(guids_in_source)
+        guids_in_db_unique = set(guids_in_db)
+        LOGGER.debug('guids in source: %s, unique guids in source: %s, '\
+                      'guids in db: %s, unique guids in db: %s', len(guids_in_source),
+                      len(guids_in_source_unique), len(guids_in_db), len(guids_in_db_unique))
+        guids_to_delete = guids_in_db_unique - guids_in_source_unique
 
         # Create a harvest object for each of them, flagged for deletion
         for guid in guids_to_delete:
@@ -126,36 +144,84 @@ class DCATdeRDFHarvester(DCATRDFHarvester):
 
         return object_ids
 
-    def amend_package(self, package, portal):
+    def _amend_package(self, package, pkg_guid, portal, harvester_name):
+        '''
+        Amend package information.
+        '''
         if 'extras' not in package:
             package['extras'] = []
 
         set_extras_field(package, EXTRA_KEY_HARVESTED_PORTAL, portal)
 
+        # ensure all resources have a license
+        for resource in package.get('resources', []):
+            if RES_EXTRA_KEY_LICENSE not in resource:
+                # pkg_guid is needed because GUID is not set in package dict
+                LOGGER.info(u'{3}: No license for resource {0} of package {1} (GUID {2}). '\
+                            u'Adding default value.'.format(
+                                resource.get('uri', ''), package.get('name', ''), pkg_guid, harvester_name)
+                           )
+                resource[RES_EXTRA_KEY_LICENSE] = self._get_fallback_license()
+
+    def _skip_datasets_without_resource(self, harvest_object, package):
+        '''
+        Checks if resources are present when configured and the dataset not already exists.
+        '''
+        if (self._get_resources_required_config(harvest_object.source.config)\
+             and not package.get('resources')):
+            skip_notice = ''
+            if not self._read_datasets_from_db(harvest_object.guid):
+                skip_notice = ' Skipping dataset.'
+            # write details to log
+            LOGGER.info(u'{0}: Resources are required, but dataset {1} (GUID {2}) has none.{3}'.format(
+                harvest_object.source.title, package.get('name', ''), harvest_object.guid, skip_notice))
+            if skip_notice:
+                return True
+        return False
+
     def import_stage(self, harvest_object):
+        '''
+        Import stage for the DCAT-AP.de harvester.
+        '''
+
+        LOGGER.debug('In DCATdeRDFHarvester import_stage')
+
         # override delete logic
         status = self._get_object_extra(harvest_object, 'status')
         if status == 'delete':
             HarvestUtils.rename_delete_dataset_with_id(harvest_object.package_id)
-            LOGGER.info('Deleted package {0} with guid {1}'.format(harvest_object.package_id,
-                                                                   harvest_object.guid))
+            LOGGER.info(u'Deleted package {0} with guid {1}'.format(harvest_object.package_id,
+                                                                    harvest_object.guid))
             return True
 
-        portal = self._get_portal_from_config(harvest_object.source.config)
-
-        # set custom field
         package = json.loads(harvest_object.content)
-        self.amend_package(package, portal)
+
+        # skip if resources are not present when configured
+        if self._skip_datasets_without_resource(harvest_object, package):
+            # do not include details in error such that they get summarized in the UI
+            self._save_object_error(
+                'Dataset has no resources, but they are required by config. Skipping.',
+                harvest_object, 'Import'
+                )
+            return False
+
+        # set custom field and perform other fixes on the data
+        portal = self._get_portal_from_config(harvest_object.source.config)
+        self._amend_package(package, harvest_object.guid, portal, harvest_object.source.title)
         harvest_object.content = json.dumps(package)
+
         import_dataset = HarvestUtils.handle_duplicates(harvest_object.content)
         if import_dataset:
             return super(DCATdeRDFHarvester, self).import_stage(harvest_object)
         else:
-            self._save_object_error('Skipping importing dataset %s, because of duplicate detection!' %
-                                    (package.get('name', '')), harvest_object, 'Import')
+            self._save_object_error('Skipping importing dataset, because of duplicate detection!',
+                                    harvest_object, 'Import')
             return False
 
     def validate_config(self, source_config):
+        '''
+        Validates additional configuration parameters for DCAT-AP.de harvester.
+        '''
         cfg = super(DCATdeRDFHarvester, self).validate_config(source_config)
 
         if cfg: