+++ ckanext/dcatde/harvesters/harvest_utils.py
@@ -8,12 +8,13 @@ For instance, they provide functions to rename datasets before they get deleted.
 import json
 import logging
 import uuid
+import pytz
 
 from ckan import model
 from ckan.model import Session, PACKAGE_NAME_MAX_LENGTH
 import ckan.plugins as p
-from ckanext.dcatde.extras import Extras
 from dateutil.parser import parse as parse_date
+from ckanext.dcatde.extras import Extras
 
 
 LOGGER = logging.getLogger(__name__)
@@ -31,8 +32,8 @@ class HarvestUtils(object):
 
     '''The class with utility functions for the harvesting process.'''
 
-    @classmethod
-    def build_context(cls):
+    @staticmethod
+    def build_context():
         """
         Builds a context dictionary.
         """
@@ -44,8 +45,8 @@ class HarvestUtils(object):
             'ignore_auth': True
         }
 
-    @classmethod
-    def create_new_name_for_deletion(cls, name):
+    @staticmethod
+    def create_new_name_for_deletion(name):
         """
         Creates new name by adding suffix "-deleted" and
         random string to the given name
@@ -54,8 +55,8 @@ class HarvestUtils(object):
         new_name = name[:NAME_MAX_LENGTH]
         return new_name + NAME_DELETED_SUFFIX + random_suffix
 
-    @classmethod
-    def rename_datasets_before_delete(cls, deprecated_package_dicts):
+    @staticmethod
+    def rename_datasets_before_delete(deprecated_package_dicts):
         """
         Renames the given packages to avoid name conflicts with
         deleted packages.
@@ -64,9 +65,9 @@ class HarvestUtils(object):
         renamed_package_ids = []
         package_update = p.toolkit.get_action('package_update')
         for package_dict in deprecated_package_dicts:
-            context = cls.build_context()
+            context = HarvestUtils.build_context()
             package_id = package_dict['id']
-            package_dict['name'] = cls.create_new_name_for_deletion(package_dict['name'])
+            package_dict['name'] = HarvestUtils.create_new_name_for_deletion(package_dict['name'])
             # Update package
             try:
                 package_update(context, package_dict)
@@ -76,8 +77,8 @@ class HarvestUtils(object):
 
         return renamed_package_ids
 
-    @classmethod
-    def delete_packages(cls, package_ids):
+    @staticmethod
+    def delete_packages(package_ids):
         """
         Deletes the packages belonging to the given package ids.
         """
@@ -85,7 +86,7 @@ class HarvestUtils(object):
         deleted_package_ids = []
         package_delete = p.toolkit.get_action('package_delete')
         for to_delete_id in package_ids:
-            context = cls.build_context()
+            context = HarvestUtils.build_context()
             try:
                 package_delete(context, {'id': to_delete_id})
                 deleted_package_ids.append(to_delete_id)
@@ -97,42 +98,43 @@ class HarvestUtils(object):
                 )
         return deleted_package_ids
 
-    @classmethod
-    def rename_delete_dataset_with_id(cls, package_id):
+    @staticmethod
+    def rename_delete_dataset_with_id(package_id):
         """
         Deletes the package with package_id. Before deletion, the package is renamed to avoid
         conflicts when adding new packages.
         """
-        context = cls.build_context()
+        context = HarvestUtils.build_context()
         harvester_package = p.toolkit.get_action('package_show')(context, {'id': package_id})
         # rename and delete the package (renamed_ids contains the current package ID at most)
-        renamed_ids = cls.rename_datasets_before_delete([harvester_package])
-        cls.delete_packages(renamed_ids)
+        renamed_ids = HarvestUtils.rename_datasets_before_delete([harvester_package])
+        HarvestUtils.delete_packages(renamed_ids)
 
-    @classmethod
-    def compare_metadata_modified(cls, remote_modified, local_modified):
+    @staticmethod
+    def compare_metadata_modified(remote_modified, local_modified):
         '''
-        Compares the modified datetimes of the metadata
+        Compares the modified datetimes of the metadata. Returns True if the remote date (first parameter)
+        is newer.
         '''
-        remote_dt = parse_date(remote_modified)
-        local_dt = parse_date(local_modified)
+        remote_dt = _parse_date(remote_modified)
+        local_dt = _parse_date(local_modified)
         if remote_dt < local_dt:
             LOGGER.debug('remote dataset precedes local dataset -> skipping.')
             return False
         elif remote_dt == local_dt:
             LOGGER.debug('remote dataset equals local dataset -> skipping.')
             return False
-        else:
-            LOGGER.debug('local dataset precedes remote dataset -> importing.')
-            # TODO do I have to delete other dataset?
-            return True
 
-    @classmethod
-    def handle_duplicates(cls, harvest_object_content):
+        LOGGER.debug('local dataset precedes remote dataset -> importing.')
+        # TODO do I have to delete other dataset?
+        return True
+
+    @staticmethod
+    def handle_duplicates(harvest_object_content):
         '''Compares new dataset with existing and checks, if a dataset should be imported.'''
 
         method_prefix = 'handle_duplicates: '
-        context = cls.build_context()
+        context = HarvestUtils.build_context()
 
         remote_dataset = json.loads(harvest_object_content)
         remote_dataset_extras = Extras(remote_dataset['extras'])
@@ -145,7 +147,7 @@ class HarvestUtils(object):
                 try:
                     data_dict = {"q": EXTRAS_KEY_DCT_IDENTIFIER + ':"' + orig_id + '"'}
                     # Add filter that local dataset guid is not equal to guid of the remote dataset
-                    if (remote_dataset_extras.key('guid')):
+                    if remote_dataset_extras.key('guid'):
                         data_dict['fq'] = '-guid:"' + remote_dataset_extras.value('guid') + '"'
                     local_search_result = p.toolkit.get_action("package_search")(context, data_dict)
                     if local_search_result['count'] == 0:
@@ -158,14 +160,25 @@ class HarvestUtils(object):
                         local_dataset = local_search_result['results'][0]
                         local_dataset_extras = Extras(local_dataset['extras'])
 
-                        # TODO : Im Zweifel das CKAN-Feld "metadata_modified" des lokalen Datensatzes nutzen,
-                        # falls modified nicht enthalten ist?
+                        # dct:modified vergleichen
                         if remote_dataset_extras.key(EXTRAS_KEY_DCT_MODIFIED) and \
                                 local_dataset_extras.key(EXTRAS_KEY_DCT_MODIFIED):
-                            return cls.compare_metadata_modified(
+                            return HarvestUtils.compare_metadata_modified(
                                 remote_dataset_extras.value(EXTRAS_KEY_DCT_MODIFIED),
                                 local_dataset_extras.value(EXTRAS_KEY_DCT_MODIFIED)
                             )
+                        elif remote_dataset_extras.key('metadata_harvested_portal') == \
+                                local_dataset_extras.key('metadata_harvested_portal'):
+                            # Deletion will be done by the harvester, because the datasets are from the same
+                            # metadata_harvested_portal (harvester)
+                            LOGGER.debug(
+                                '%sFound duplicate entry with the value "%s" in field "identifier", but ' \
+                                'remote and/or local dataset does not contain a modified date. ' \
+                                'No comparison possible! But the datasets are from the same harvester. ' \
+                                'Import accepted for %s.',
+                                method_prefix, orig_id, remote_dataset_name)
+                            return True
+                        # Kein Vergleich möglich, wenn nicht beide Datensätze ein modified date enthalten
                         else:
                             LOGGER.info(
                                 '%sFound duplicate entry with the value "%s" in field "identifier", but ' \
@@ -175,7 +188,7 @@ class HarvestUtils(object):
                     else:
                         LOGGER.info('%sFound multiple duplicates with the value "%s" in field ' \
                             '"identifier". -> Skipping import for %s!', method_prefix, orig_id,
-                            remote_dataset_name)
+                                    remote_dataset_name)
                 except Exception as exception:
                     LOGGER.error(exception)
             else:
@@ -188,3 +201,13 @@ class HarvestUtils(object):
             return True
 
         return False
+
+
+def _parse_date(date_string):
+    '''
+    Parse a date string to a date object and add the time zone UTC if no time zone info exists.
+    '''
+    date_obj = parse_date(date_string)
+    if date_obj.tzinfo is None:
+        date_obj = date_obj.replace(tzinfo=pytz.UTC)
+    return date_obj